Document how to use UTF-8 encoding.
log4net uses the system default encoding, which is typically Windows-1252 on .NET Framework 4.8. (.NET Core will use UTF-8 by default.) If the logs are forwarded to New Relic, the log forwarder expects them to be encoded in UTF-8 (without a BOM).

This change updates the documentation to help users create a better configuration for that platform.