Explicitly use UTF-8 encoding in examples (#88)
* Explicitly use UTF-8 encoding in examples.

log4net uses the system default encoding, which is typically Windows-1252 on .NET Framework 4.8. If the logs are forwarded to New Relic, the log forwarder expects them to be encoded in UTF-8 (without a BOM).

This change updates the documentation to help users create a better configuration by default.

* Add namespace to load encoding type class from

* Document how to use UTF-8 encoding.

log4net uses the system default encoding, which is typically Windows-1252 on .NET Framework 4.8. (.NET Core will use UTF-8 by default.) If the logs are forwarded to New Relic, the log forwarder expects them to be encoded in UTF-8 (without a BOM).

This change updates the documentation to help users create a better configuration for that platform.

* Add commented-out encoding element to log4net.config

* Revert to netcoreapp3.0 to bypass Synk problem

Co-authored-by: Alex Hemsath <ahemsath@newrelic.com>